Issueboard-Automation/requirements.txt
python-gitlab
slack-sdk
pyyaml